Cleanup the Python packaging code
Also removed the prebuild wheels as wheel as CI autogenerated now
