+++ Manylinux2014_Compliant_Source/pkg/IR2Vec/README.md
@@ -1,0 +1,1 @@
+given input .ll/.bc generates corresponding IR2Vec embeddings in a file or on stdout
