+++ Manylinux2014_Compliant_Source/pkg/IR2Vec/__init__.py
@@ -1,0 +1,5 @@
+### we have to find package before importing it else export LD_Lib path and LIB_path will not work.
+from . import preparation
+from IR2Vec_pkg.core import *
+
+setSeedEmbdPath(preparation.install_loc_pkg)
