+++ Manylinux2014_Compliant_Source/pkg/IR2Vec/preparation.py
@@ -1,0 +1,4 @@
+import pathlib as pl
+
+install_loc = pl.Path(__file__).parent.resolve()
+install_loc_pkg = str(install_loc)
