+++ Manylinux2014_Compliant_Source/pkg/Manifest.in
@@ -1,2 +1,1 @@
 include ./seedEmbeddingVocab-300-llvm12.txt
-recursive-include ./IR2Vec_include/*
