+++ Manylinux2014_Compliant_Source/pkg/setup.py
@@ -1,41 +1,81 @@
 from setuptools import setup, Extension
 
 import subprocess as sp
+import pathlib as pl
+
+import re
+
+version_regex = re.compile(r"^project\(ir2vec VERSION (?P<version>[^)]+)\)$")
+llvm_libs_regex = re.compile(
+    r"^llvm_map_components_to_libnames\(llvm_libs (?P<libs>[^)]+)\)$"
+)
+
+LLVM_LIBS = []
+VERSION = ""
+DESCRIPTION = ""
+
+with (pl.Path(__file__).parents[2] / "src" / "CMakeLists.txt").open() as f:
+    for line in f:
+        if not VERSION:
+            vmatch = version_regex.match(line)  # Not using walrus because Python3.6
+            if vmatch:
+                VERSION = vmatch.group("version")
+                continue
+        if not LLVM_LIBS:
+            libmatch = llvm_libs_regex.match(line)
+            if libmatch:
+                LLVM_LIBS = libmatch.group("libs").split()
+                continue
+        if VERSION and LLVM_LIBS:
+            break
+
+with (pl.Path(__file__).parent / "IR2Vec" / "README.md").open() as f:
+    DESCRIPTION = f.read()
+
 
 def get_llvm_files():
-    libs = "support core irreader analysis TransformUtils".split()
-    out = sp.run(['llvm-config', '--libfiles'] + libs, stdout=sp.PIPE)
-    files = out.stdout.decode('utf8').split()
+    out = sp.run(["llvm-config", "--libfiles"] + LLVM_LIBS, stdout=sp.PIPE)
+    files = out.stdout.decode("utf8").split()
     return files
 
 
 IR2Vec_core = Extension(
-    "IR2Vec_pkg.core",
-    sources=["IR2Vec_pkg/core.cpp"],
-    include_dirs=[
-        "./IR2Vec_pkg",
-        "./IR2Vec_pkg/IR2Vec_include",
-    ],  # list of directories to search for C/C++ header files (in Unix form for portability)
+    "IR2Vec.core",
+    sources=["IR2Vec/core.cpp"],
+    include_dirs=["./IR2Vec"],
     libraries=["z"],
     extra_objects=["/usr/local/lib/libIR2Vec.a"] + get_llvm_files(),
     extra_compile_args=["-v"],
 )
 
 setup(
-    name="IR2Vec_pkg",
+    name="IR2Vec",
     author="Shikhar Jain",
     author_email="cs22mtech02002@iith.ac.in",
-    version="1.0.0",
+    version=VERSION,
     description="given input .ll/.bc generates corresponding IR2Vec embeddings in a file or on stdout",
-    long_description="given input .ll/.bc generates corresponding IR2Vec embeddings in a file or on stdout",
-    ext_modules=[IR2Vec_core],  # A list of Python extensions to be built
-    packages=["IR2Vec_pkg"],  # A list of Python packages that distutils will manipulate
-    # package_data={'':['*.so*','./IR2Vec_include/*','./llvm/*','./llvm-c/*','seedEmbeddingVocab-300-llvm12.txt']},
-    package_data={
-        "": [
-            "./IR2Vec_include/*",
-            "seedEmbeddingVocab-300-llvm12.txt",
-        ]
-    },
+    long_description=DESCRIPTION,
+    long_description_content_type="text/markdown",
+    url="https://github.com/IITH-Compilers/IR2Vec",
+    license="BSD 4-Clause",
+    classifiers=[
+        "Development Status :: 5 - Production/Stable",
+        "Intended Audience :: Developers",
+        "Intended Audience :: Science/Research",
+        "License :: OSI Approved :: BSD License",
+        "Programming Language :: C++",
+        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3.6",
+        "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
+        "Topic :: Scientific/Engineering :: Artificial Intelligence",
+        "Topic :: Software Development :: Compilers",
+    ],
+    ext_modules=[IR2Vec_core],
+    packages=["IR2Vec"],
+    package_data={"": ["seedEmbeddingVocab-300-llvm12.txt"]},
     include_package_data=True,
 )
